requirements.txt
Flask
SQLAlchemy
Flask-Migrate
Werkzeug
Pillow
